remove filtered-comparisons mixin file, move to lib
