+++ app/lib/filtered-comparisons.js
@@ -1,4 +1,34 @@
-import Object, {get, set} from '@ember/object';
+import Object, {computed, get, set} from '@ember/object';
+import {alias, or, filterBy} from '@ember/object/computed';
+
+export default Object.extend({
+  activeBrowser: null,
+  snapshot: null,
+  snapshotSelectedWidth: null,
+
+  _comparisons: alias('snapshot.comparisons'),
+
+  comparisons: computed('_comparisons.@each.browser', 'activeBrowser', function() {
+    return computeComparisonsForBrowser(get(this, '_comparisons'), get(this, 'activeBrowser'));
+  }),
+
+  comparisonForWidth: computed('comparisons.@each.width', 'snapshotSelectedWidth', function() {
+    const width = get(this, 'snapshotSelectedWidth') || get(this, 'defaultWidth');
+    return computeComparisonForWidth(get(this, 'comparisons'), width);
+  }),
+
+  widestComparisonForBrowser: computed('comparisons.@each.width', function() {
+    return computeWidestComparison(get(this, 'comparisons'));
+  }),
+
+  widestComparisonWithDiff: computed('comparisons.@each.width', function() {
+    return computeWidestComparisonWithDiff(get(this, 'comparisons'));
+  }),
+
+  selectedComparison: or('comparisonForWidth', 'widestComparisonForBrowser'),
+  defaultWidth: or('widestComparisonWithDiff.width', 'widestComparisonForBrowser.width'),
+  comparisonsWithDiffs: filterBy('comparisons', 'isDifferent'),
+});
 
 export function computeWidestComparison(comparisons) {
   return get(comparisons.sortBy('width'), 'lastObject');
